requirements.txt
alembic==1.1.0
blinker==1.4
click==7.0
flask==1.1.1
flask-login==0.4.1
flask-mail==0.9.1
flask-migrate==2.5.2
flask-sqlalchemy==2.4.0
itsdangerous==1.1.0
jinja2==2.10.1
mako==1.1.0
markupsafe==1.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
sqlalchemy==1.3.8
survey-flask-app==0.0.1
werkzeug==0.16.0
